working

+++ Assets/Script/Models/DroneVelocityModel.cs
@@ -6,8 +6,8 @@ namespace Server.Models
     public class DroneVelocityModel
     {
         public string droneId { get; set; }
-        public int velocityX { get; set; }
-        public int velocityY { get; set; }
+        public float velocityX { get; set; }
+        public float velocityY { get; set; }
         public int rate { get; set; }
     }
 }
